Refactor handling of one object in one file
