+++ libzazen/keyboard.c
@@ -1,0 +1,16 @@
+#include "keyboard.h"
+
+#include "util.h"
+
+struct zazen_keyboard* zazen_keyboard_create(struct zazen_seat* seat)
+{
+  UNUSED(seat);
+  // TODO: create keyboard
+  return NULL;
+}
+
+void zazen_keyboard_destroy(struct zazen_keyboard* keyboard)
+{
+  UNUSED(keyboard);
+  // TODO: destroy keyboard
+}
